search fix + 5 latest questions
