+++ server.py
@@ -3,7 +3,7 @@ from question import question_api
 from answer import answer_api
 from comment import comment_api
 from connection import QUESTION, IMAGE_DATA
-from data_handler import read_all_data_from_db, sort_questions_by_order, search_database_by_string
+from data_handler import read_all_data_from_db, sort_questions_by_order, search_database_by_string, top_five_latest_question
 
 app = Flask(__name__, template_folder='templates', static_folder='static', )
 app.register_blueprint(question_api)
@@ -16,12 +16,12 @@ def route_image(filename):
 
 @app.route("/")
 def route_home():
-    return render_template("index.html")
+    questions = top_five_latest_question()
+    return render_template("index.html",questions=questions)
 
 @app.route("/list", methods=["GET"])
 def route_list():
     questions = read_all_data_from_db(QUESTION)
-    # Sort list page
     if request.args.get("search") != None:
         return redirect(url_for("route_search",search=request.args.get("search")))
     if request.args.get("order_by") != None and request.args.get("order_direction") != None:
@@ -33,7 +33,6 @@ def route_list():
 @app.route("/search", methods=["GET"])
 def route_search():
     questions = search_database_by_string(request.args.get('search'))
-    print(questions)
     return render_template("search.html", questions=questions)
 
 if __name__ == "__main__":
